Gate non-wasm32 winit features behind flags.
